fix(tag): Roll back to working tag

+++ src/smbase/crc.h
@@ -4,6 +4,6 @@
 #ifndef __CRC_H
 #define __CRC_H
 
-uint32_t crc32(unsigned char const *data, int length);
+unsigned long crc32(unsigned char const *data, int length);
 
 #endif // __CRC_H
